add queue in galleryController , on doing deleted events

+++ app/Http/Controllers/ScoresController.php
@@ -16,11 +16,11 @@ class ScoresController extends Controller
   public function __construct()
   {
     $this->middleware('auth', [
-        'only' => ['user_scored', 'destroy']
+        'only' => ['user_recent_scored', 'destroy']
     ]);
   }
 
-  public function user_scored(User $user)
+  public function user_recent_scored(User $user)
   {
     $scores = Score::where('user_id', $user->id)->get();
     return view('index.scores', compact('scores'));
@@ -28,6 +28,8 @@ class ScoresController extends Controller
 
   public function destroy(Score $score)
   {
+    $this->authorize('update', Auth::user());
+    
     $score->delete();
     return redirect()->route('scores.user_scored', Auth::id());
   }
